windows_receiver/requirements.txt
websockets==12.0
sounddevice==0.4.7
pyinstaller==6.11.1
pystray==0.19.5
Pillow==10.4.0
